+++ backend/src/main.rs
@@ -185,7 +185,7 @@ async fn start_in_cli_mode(cfg: Arc<AppConfig>, targets: Arc<ProcessTargets>) {
         error!("Failed to build client {err}");
         reqwest::Client::new()
     });
-    playlist::exec_processing(Arc::new(client), cfg, targets, None, None).await;
+    playlist::exec_processing(&client, cfg, targets, None, None).await;
 }
 
 async fn start_in_server_mode(cfg: Arc<AppConfig>, targets: Arc<ProcessTargets>) {