+++ backend/src/utils/json_utils.rs
@@ -1,4 +1,4 @@
-use crate::utils::file_reader;
+use crate::utils::{file_reader, file_writer};
 use serde::Serialize;
 use serde_json::Value;
 use std::collections::{HashMap, HashSet};
@@ -49,8 +49,23 @@ where
     T: Serialize,
 {
     let file = std::fs::File::create(path)?;
-    let mut buf_writer = std::io::BufWriter::new(file);
+    let mut buf_writer = file_writer(file);
     serde_json::to_writer(&mut buf_writer, value)?;
     buf_writer.flush()?;
     buf_writer.into_inner()?.sync_all()
 }
+
+// pub async fn is_valid_json_file(path: &str) -> std::io::Result<bool> {
+//     if let Ok(file) = tokio::fs::File::open(path).await {
+//         let reader = async_file_reader(file);
+//         let stream = serde_json::Deserializer::from_reader(reader).into_iter::<serde_json::Value>();
+//         for item in stream {
+//             if item.is_err() {
+//                 return Ok(false);
+//             }
+//         }
+//         Ok(true)
+//     } else {
+//         Ok(false)
+//     }
+// }